# Avoid repeatedly copying node when evicting jobs