--- a/internal/scheduler/preempting_queue_scheduler_test.go
+++ b/internal/scheduler/preempting_queue_scheduler_test.go
@@ -1352,13 +1352,7 @@ func BenchmarkPreemptingQueueScheduler(b *testing.B) {
 		MaxPriorityFactor int
 	}{
 		"1 node 1 queue 320 jobs": {
-			SchedulingConfig: testfixtures.WithNodeOversubscriptionEvictionProbabilityConfig(
-				0,
-				testfixtures.WithNodeEvictionProbabilityConfig(
-					0.1,
-					testfixtures.TestSchedulingConfig(),
-				),
-			),
+			SchedulingConfig:  testfixtures.TestSchedulingConfig(),
 			Nodes:             testfixtures.N32CpuNodes(1, testfixtures.TestPriorities),
 			JobFunc:           testfixtures.N1CpuJobs,
 			NumQueues:         1,
@@ -1367,10 +1361,7 @@ func BenchmarkPreemptingQueueScheduler(b *testing.B) {
 			MaxPriorityFactor: 1,
 		},
 		"10 nodes 1 queue 3200 jobs": {
-			SchedulingConfig: testfixtures.WithNodeEvictionProbabilityConfig(
-				0.1,
-				testfixtures.TestSchedulingConfig(),
-			),
+			SchedulingConfig:  testfixtures.TestSchedulingConfig(),
 			Nodes:             testfixtures.N32CpuNodes(10, testfixtures.TestPriorities),
 			JobFunc:           testfixtures.N1CpuJobs,
 			NumQueues:         1,
@@ -1379,10 +1370,7 @@ func BenchmarkPreemptingQueueScheduler(b *testing.B) {
 			MaxPriorityFactor: 1,
 		},
 		"10 nodes 10 queues 3200 jobs": {
-			SchedulingConfig: testfixtures.WithNodeEvictionProbabilityConfig(
-				0.1,
-				testfixtures.TestSchedulingConfig(),
-			),
+			SchedulingConfig:  testfixtures.TestSchedulingConfig(),
 			Nodes:             testfixtures.N32CpuNodes(10, testfixtures.TestPriorities),
 			JobFunc:           testfixtures.N1CpuJobs,
 			NumQueues:         10,
@@ -1391,10 +1379,7 @@ func BenchmarkPreemptingQueueScheduler(b *testing.B) {
 			MaxPriorityFactor: 1,
 		},
 		"100 nodes 1 queue 32000 jobs": {
-			SchedulingConfig: testfixtures.WithNodeEvictionProbabilityConfig(
-				0.1,
-				testfixtures.TestSchedulingConfig(),
-			),
+			SchedulingConfig:  testfixtures.TestSchedulingConfig(),
 			Nodes:             testfixtures.N32CpuNodes(100, testfixtures.TestPriorities),
 			JobFunc:           testfixtures.N1CpuJobs,
 			NumQueues:         1,
@@ -1403,10 +1388,7 @@ func BenchmarkPreemptingQueueScheduler(b *testing.B) {
 			MaxPriorityFactor: 1,
 		},
 		"1000 nodes 1 queue 320000 jobs": {
-			SchedulingConfig: testfixtures.WithNodeEvictionProbabilityConfig(
-				0.1,
-				testfixtures.TestSchedulingConfig(),
-			),
+			SchedulingConfig:  testfixtures.TestSchedulingConfig(),
 			Nodes:             testfixtures.N32CpuNodes(1000, testfixtures.TestPriorities),
 			JobFunc:           testfixtures.N1CpuJobs,
 			NumQueues:         1,